pyproject.toml
[tool.poetry]
name = "essence-demo"
version = "0.1.0"
description = ""
authors = ["Samuil Plamenov <samuil.plamenov@harpercollins.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
Flask-Login = "^0.6.1"
oauthlib = "^3.2.0"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
alpaca-trade-api = "^2.2.0"
plotly = "^5.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
